requirements.txt
grpcio
grpcio-tools
Flask-SQLAlchemy
PyJWT